Fix corner case bug in day24 solution
I had a comment in the code - it turned out that it was my mistake.
Note: in puzzle description for example data (data_small.txt) there is an answer not matching what my code produces:
after 100 days: 2208, in my code it is 2136 (the first 9 days are the same, difference is from day 10: my 39, example 37).
The program works for full dataset - it produced correct answer 3964.
I spent some time debugging if there is a bug here despite correct answer, but I could not find one.
Still I am not sure if there was a bug in the example answer or if my solution does not work for some corner cases.

Indeed it was a corner case: I used the same sets src, dst when iterating in tiles_game_of_life.
I knew that data remains in next iterations, so I always discarded element from the set when it
was a white tile. But I checked positions only within the borders of considered space.
It worked most of the time, because the space was getting larger.
But if it was smaller in some iteration - then I would not remove redundant data.
Fixed in a simple way - instead of swapping sets in each iteration, create new black_tiles set in each one.